# 4.x: Improve Buffer via dedicated exact/skip implementations